requirements.txt
# Core
numpy==1.24.2
scipy==1.10.1
pandas==1.5.3
matplotlib==3.7.3
seaborn==0.12.2
Pillow==9.4.0
PyYAML==6.0
tqdm==4.66.1
scikit-learn==1.3.0
joblib==1.3.2

# RL & Simulation
gym==0.26.2
ray==2.3.1
cloudpickle==2.2.1
sumolib==1.16.0
traci==1.16.0

# Deep Learning Utilities
einops==0.6.0
kmeans-pytorch==0.3
tensorboard==2.13.0

# Logging
wandb==0.14.2

# Other
requests==2.28.2
click==8.1.3
packaging==23.0
filelock==3.11.0
